+++ playback/src/player.rs
@@ -138,12 +138,9 @@ enum PlayerCommand {
         track: bool,
     },
     EmitAutoPlayChangedEvent(bool),
-    EmitQueueChangedEvent(String),
-    EmitContextLoadedEvent {
-        context_uri: String,
-        current_track: Option<(String, String)>, // (uri, provider)
-        next_tracks: Vec<(String, String)>,      // (uri, provider)
-        prev_tracks: Vec<(String, String)>,      // (uri, provider)
+    EmitSetQueueEvent {
+        next_tracks: Vec<(String, String)>, // (uri, provider)
+        prev_tracks: Vec<(String, String)>, // (uri, provider)
     },
 }
 
@@ -230,6 +227,10 @@ pub enum PlayerEvent {
     },
     TrackChanged {
         audio_item: Box<AudioItem>,
+    },
+    SetQueue {
+        next_tracks: Vec<(String, String)>, // (uri, provider)
+        prev_tracks: Vec<(String, String)>, // (uri, provider)
     },
     SessionConnected {
         connection_id: String,
@@ -665,16 +666,12 @@ impl Player {
         self.command(PlayerCommand::EmitAutoPlayChangedEvent(auto_play));
     }
 
-    pub fn emit_context_loaded_event(
+    pub fn emit_set_queue_event(
         &self,
-        context_uri: String,
-        current_track: Option<(String, String)>,
         next_tracks: Vec<(String, String)>,
         prev_tracks: Vec<(String, String)>,
     ) {
-        self.command(PlayerCommand::EmitContextLoadedEvent {
-            context_uri,
-            current_track,
+        self.command(PlayerCommand::EmitSetQueueEvent {
             next_tracks,
             prev_tracks,
         });
@@ -2379,8 +2376,14 @@ impl PlayerInternal {
                 self.auto_normalise_as_album = setting
             }
 
-            PlayerCommand::EmitAddedToQueueEvent(track_id) => {
-                self.send_event(PlayerEvent::AddedToQueue { track_id })
+            PlayerCommand::EmitSetQueueEvent {
+                next_tracks,
+                prev_tracks,
+            } => {
+                self.send_event(PlayerEvent::SetQueue {
+                    next_tracks,
+                    prev_tracks,
+                })
             }
 
             PlayerCommand::EmitFilterExplicitContentChangedEvent(filter) => {
@@ -2584,9 +2587,13 @@ impl fmt::Debug for PlayerCommand {
                 .debug_tuple("EmitAutoPlayChangedEvent")
                 .field(&auto_play)
                 .finish(),
-            PlayerCommand::EmitContextLoadedEvent { context_uri, .. } => f
-                .debug_tuple("EmitContextLoadedEvent")
-                .field(&context_uri)
+            PlayerCommand::EmitSetQueueEvent {
+                next_tracks,
+                prev_tracks,
+            } => f
+                .debug_tuple("EmitSetQueueEvent")
+                .field(&next_tracks.len())
+                .field(&prev_tracks.len())
                 .finish(),
         }
     }