+++ src/player_event_handler.rs
@@ -234,17 +234,11 @@ impl EventHandler {
                                 );
                                 env_vars.insert("FILTER", filter.to_string());
                             }
-                            PlayerEvent::ContextLoaded {
-                                context_uri,
-                                current_track,
+                            PlayerEvent::SetQueue {
                                 next_tracks,
                                 prev_tracks,
                             } => {
-                                env_vars.insert("PLAYER_EVENT", "context_loaded".to_string());
-                                env_vars.insert("CONTEXT_URI", context_uri);
-                                if let Some((uri, provider)) = current_track {
-                                    env_vars.insert("CURRENT_TRACK", format!("{uri}\t{provider}"));
-                                }
+                                env_vars.insert("PLAYER_EVENT", "set_queue".to_string());
                                 env_vars.insert(
                                     "NEXT_TRACKS",
                                     next_tracks