+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [playback] Add `PlayerEvent::ContextLoaded` event to notify when a context (playlist, album, etc.) is loaded with its track list
+- [playback] Add `PlayerEvent::SetQueue` event to notify when the play queue changes
 
 ### Changed